codet5-base.commons-lang.mid-2294.idx-16390.3.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableByte.java
@@ -309,7 +309,7 @@ public class MutableByte extends Number implements Comparable<MutableByte>, Muta
      */
     @Override
     public float floatValue() {
-        return value;
+    return (float) value;
     }
 
     /**
